the project is now reset up without having the weird login auth that I was using before with tempo-goals but idk why it is acting weird
when a user logs in, everything is fine but if the user reloads the brower the data will go away and errors in the console come up. I'm guessing this is becuse of how App.vue is the compent that checks for the user now but I think it is best if it is done there? how can I pass all of the info for should it be made in UserGoals.vue